moved userProfile endpoint in with user.js
