+++ backend/src/app.js
@@ -28,7 +28,7 @@ const organizationRouter = require('./routes/organization');
 const publicRouter = require('./routes/public');
 const configRouter = require('./routes/config');
 
-const userprofileRouter = require('./routes/userprofile');
+//const userprofileRouter = require('./routes/userprofile');
 
 
 const promMid = require('express-prometheus-middleware');
@@ -168,7 +168,7 @@ apiRouter.use('/facility', facilityRouter);
 apiRouter.use('/organization', organizationRouter);
 apiRouter.use('/public', publicRouter);
 apiRouter.use('/config',configRouter);
-apiRouter.use('/profile', userprofileRouter);
+//apiRouter.use('/profile', userprofileRouter);
